# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MCP Multi-Agent Chat System API
  description: |
    A powerful multi-agent AI system for collaborative software development, featuring 
    specialized agents for architecture, coding, testing, security, and DevOps, with 
    advanced NLU capabilities and intelligent task orchestration.
  version: 2.0.0
  contact:
    name: API Support
    url: https://github.com/anthropics/claude-code/issues
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5006/api
    description: Development server
  - url: https://api.example.com/api/v2
    description: Production server (v2)
  - url: https://api.example.com/api/v1
    description: Production server (v1 - deprecated)
security:
  - ApiKeyAuth: []
...
paths:
...
  /tasks/{task_id}/status:
    get:
      tags:
        - Tasks
      summary: Get task status
      description: Retrieve the current status and progress of a task
      operationId: getTaskStatus
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatus'
  /tasks:
    get:
      tags:
        - Tasks
      summary: List tasks
      description: Get a paginated list of tasks
      operationId: listTasks
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed, failed]
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
...
  /audit/statistics:
    get:
      tags:
        - Audit
      summary: Get audit statistics
      description: Retrieve audit system statistics and metrics
      operationId: getAuditStatistics
      x-available-since: "2.0.0"
      parameters:
        - name: agent_id
          in: query
          schema:
            type: string
          description: Filter by specific agent
        - name: start_date
          in: query
          schema:
            type: string
            format: date
          description: Start date for statistics
        - name: end_date
          in: query
          schema:
            type: string
            format: date
          description: End date for statistics
      responses:
        '200':
          description: Audit statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditStatisticsResponse'
...
  /audit/level:
    post:
      tags:
        - Audit
      summary: Set audit level
      description: Configure the level of detail for audit logging
      operationId: setAuditLevel
      x-available-since: "2.0.0"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuditLevelRequest'
      responses:
        '200':
          description: Audit level updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
...
    TaskStatus:
      type: object
      properties:
        task_id:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        result:
          type: object
          properties:
            success:
              type: boolean
            summary:
              type: string
    TaskListResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              title:
                type: string
              status:
                type: string
              created_at:
                type: string
                format: date-time
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    AuditStatisticsResponse:
      type: object
      properties:
        success:
          type: boolean
        statistics:
          type: object
          properties:
            total_records:
              type: integer
            success_rate:
              type: number
              format: float
            agents:
              type: object
              additionalProperties:
                type: object
                properties:
                  tasks:
                    type: integer
                  success_rate:
                    type: number
                  avg_duration:
                    type: number
            by_intent:
              type: object
              additionalProperties:
                type: integer
...
    AuditLevelRequest:
      type: object
      required:
        - level
      properties:
        level:
          type: string
          enum: [minimal, standard, detailed, debug]
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
...